+++ app/__init__.py
@@ -11,5 +11,6 @@ migrate = Migrate(app, db)
 login = LoginManager(app)
 login.login_view = 'login'
 login.login_message_category = 'info'
+login.refresh_view = 'login'
 
 from app import routes, models, errors